requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.4.1
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
exceptiongroup==1.2.0
flake8==7.0.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.0
pluggy==1.4.0
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.6.0
pytest==8.0.0
pytest-cov==4.1.0
pytest-django==4.8.0
pytz==2023.2
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.2
uritemplate==4.1.1
urllib3==1.26.15
